feat(SearchViewModel): New view model to hold the async search logic
