+++ src/gui4/macOS/kDrive/UI/Views/MainWindow/Search/SearchViewModel.swift
@@ -1,0 +1,65 @@
+/*
+ * Infomaniak kDrive - Desktop
+ * Copyright (C) 2023-2025 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Combine
+import kDriveCore
+import kDriveCoreUI
+
+@MainActor
+final class SearchViewModel: ObservableObject {
+    @Published var searchText = ""
+    @Published private(set) var searchResults: [UIFileResponse] = []
+    @Published private(set) var isSearching = false
+
+    private let syncDbId: Int32
+    private var bindStore = Set<AnyCancellable>()
+
+    init(syncDbId: Int32) {
+        self.syncDbId = syncDbId
+        setupSearchSubscription()
+    }
+
+    private func setupSearchSubscription() {
+        $searchText
+            .throttle(for: .milliseconds(250), scheduler: DispatchQueue.main, latest: true)
+            .sink { [weak self] query in
+                self?.performSearch(query: query)
+            }
+            .store(in: &bindStore)
+    }
+
+    private func performSearch(query: String) {
+        guard !query.isEmpty else {
+            searchResults = []
+            isSearching = false
+            return
+        }
+
+        isSearching = true
+        Task {
+            do {
+                let results = try await DriveJobs().driveSearch(syncDbId: syncDbId, searchString: query)
+                searchResults = results.map { UIFileResponse(fileResponse: $0) }
+                isSearching = false
+            } catch {
+                searchResults = []
+                isSearching = false
+            }
+        }
+    }
+}
